Replace circular-dependency-plugin with eslint-plugin-import/no-cycle rule
This is a check that is better done during lint in CI rather than slowing down our build time. We were already using the eslint-plugin-import package so this change does not introduce any new packages and lets us remove the circular-dependency-plugin package.
