+++ packages/studio-base/package.json
@@ -61,7 +61,6 @@
     "@testing-library/react-hooks": "8.0.1",
     "@testing-library/user-event": "14.4.3",
     "@types/base16": "^1.0.2",
-    "@types/circular-dependency-plugin": "5.0.5",
     "@types/cytoscape": "^3.19.9",
     "@types/dat.gui": "^0.7.7",
     "@types/foxglove__roslibjs": "workspace:*",
@@ -97,7 +96,6 @@
     "chartjs-plugin-annotation": "2.2.1",
     "chartjs-plugin-datalabels": "2.2.0",
     "chromatic": "6.17.3",
-    "circular-dependency-plugin": "5.2.2",
     "comlink": "4.4.1",
     "crypto-browserify": "3.12.0",
     "css-loader": "6.7.3",
